requirement.txt
asgiref==3.2.10
Django==3.1
djangorestframework==3.11.0
djangorestframework-simplejwt==5.1.0
install==1.3.5
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
